restructured stats handlers
